requirements.txt
bluepy==1.0.5
pycrypto==2.6.1
paho-mqtt
